Backend/requirements.txt
# requirements.txt
flask
flask-cors
pymongo
face_recognition
pillow
numpy
pyjwt
werkzeug